Backend/requirements.txt
django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.5.0
dj-database-url==2.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
mailjet-rest==1.5.1